backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
anthropic>=0.39.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
apscheduler==3.10.4

# Google integrations (optional - install if using Gmail/Sheets)
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0
gspread>=5.12.0
